+++ src/main/java/org/example/interfaces/PersonaService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface PersonaService {
     List<Persona> findAll();
-    Persona findByid(int id);
+    Persona findByid(long id);
     void save(Persona persona);
     void update(Persona persona);
-    void delete(int id);
+    void delete(long id);
 }
